+++ src/main/java/com/rshu/mppz/data/customer/CustomerRepository.java
@@ -1,4 +1,6 @@
-package com.rshu.mppz;
+package com.rshu.mppz.data.customer;
+
+import com.rshu.mppz.data.customer.model.Customer;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
@@ -10,17 +12,17 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collection;
 
-public interface OrderRepository extends Repository<Order, Integer> {
+public interface CustomerRepository extends Repository<Customer, Integer> {
 
 	@Transactional(readOnly = true)
-	Collection<Order> findAll() throws DataAccessException;
+	Collection<Customer> findAll() throws DataAccessException;
 
 	@Transactional(readOnly = true)
 	Page<Customer> findAll(Pageable pageable) throws DataAccessException;
 
-	@Query("SELECT order FROM Order order WHERE order.orderId =:id")
+	@Query("SELECT customer FROM Customer customer WHERE customer.customerId =:id")
 	@Transactional(readOnly = true)
-	Order findById(@Param("id") Integer id);
+	Customer findById(@Param("id") Integer id);
 
-    void save(Order order);
+	void save(Customer customer);
 }
