+++ src/main/java/com/rshu/mppz/data/order/OrderRepository.java
@@ -1,4 +1,7 @@
-package com.rshu.mppz;
+package com.rshu.mppz.data.order;
+
+import com.rshu.mppz.data.order.model.Order;
+import com.rshu.mppz.data.customer.model.Customer;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
@@ -10,17 +13,17 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collection;
 
-public interface CustomerRepository extends Repository<Customer, Integer> {
+public interface OrderRepository extends Repository<Order, Integer> {
 
 	@Transactional(readOnly = true)
-	Collection<Customer> findAll() throws DataAccessException;
+	Collection<Order> findAll() throws DataAccessException;
 
 	@Transactional(readOnly = true)
 	Page<Customer> findAll(Pageable pageable) throws DataAccessException;
 
-	@Query("SELECT customer FROM Customer customer WHERE customer.customerId =:id")
+	@Query("SELECT order FROM Order order WHERE order.orderId =:id")
 	@Transactional(readOnly = true)
-	Customer findById(@Param("id") Integer id);
+	Order findById(@Param("id") Integer id);
 
-	void save(Customer customer);
+    void save(Order order);
 }
