+++ src/main/java/com/rshu/mppz/data/product/ProductRepository.java
@@ -1,14 +1,13 @@
-package com.rshu.mppz;
+package com.rshu.mppz.data.product;
+
+import com.rshu.mppz.data.product.model.Product;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
-import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collection;
 import java.util.List;
 
 public interface ProductRepository extends Repository<Product, Integer> {
