+++ src/main/java/com/rshu/mppz/data/utils/ListStringConvertor.java
@@ -1,4 +1,4 @@
-package com.rshu.mppz;
+package com.rshu.mppz.data.utils;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.core.JsonProcessingException;
