+++ src/main/java/com/rshu/mppz/domain/customer/Customers.java
@@ -1,7 +1,9 @@
-package com.rshu.mppz;
+package com.rshu.mppz.domain.customer;
 
 //import javax.xml.bind.annotation.XmlElement;
 //import javax.xml.bind.annotation.XmlRootElement;
+import com.rshu.mppz.data.customer.model.Customer;
+
 import java.util.ArrayList;
 import java.util.List;
